// src/app/Models/AdminDashboardDetailUI.ts
export interface AdminDashboardDetailUI {
  company_Id: number
  company_Name: string
  companyShortName: string
  lot_Number: number
  headCount: number
  ctc: number
  netPay: number
  createdOn: string
  allottedDateTime?: string
  assignedTo?: string
  estimateTime: any
  qC_Verified_DateTime?: string
  score: string
  reportingManager?: string
  invoiceGenerated: string
  invoiceGeneratedDate: string
  salaryPayout: string
  customer_Confirmation_Status: any
  customer_Confirmation_DateTime: any
  process_Category?: string
  payroll_Input_Type: string
  company_Code: string
  pay_period: string
  pay_Period_Id: number
}